Admin->Users endpoints (#74)
This is the first endpoint that does not follow the URL path
at all. For Admin functions, I think it will be a common thing
because they are not well structured on server-side, but
on client-side we can do it better and organize them.

resolves #58

+++ services/federation/instances.go
@@ -15,7 +15,7 @@ type InstancesRequest struct {
 	Subscribing   bool   `json:"subscribing"`
 	Publishing    bool   `json:"publishing"`
 	Limit         uint   `json:"limit"`
-	Offset        uint   `json:"offset"`
+	Offset        uint64 `json:"offset"`
 	Sort          string `json:"sort"`
 }
